requirements.txt
# Core ML libraries
lightgbm>=4.1.0
scikit-learn>=1.4.0
shap>=0.44.0
treelite>=3.9.0
treelite_runtime>=3.9.0

# Data processing
pandas>=2.1.3
numpy>=1.26.2

# API and web
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Utilities
requests>=2.31.0
structlog>=23.2.0
python-dotenv>=1.0.0
sentry-sdk>=1.39.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
httpx>=0.25.2  # For FastAPI testing

# Monitoring (optional)
prometheus-fastapi-instrumentator>=6.1.0 